Extract BaseTrie
